# backend/app/api/routes/providers.py
"""POST /api/providers, GET /api/providers, PATCH /api/providers/{id}.

Ключ шифруется при записи, не возвращается в ответах (AGENTS.md §14).
Access control: capability "manage_providers" (только admin через "*" в seed presets).
"""

from __future__ import annotations

from fastapi import APIRouter, Depends, Request
from sqlalchemy import func, select, update
from sqlalchemy.exc import IntegrityError
from sqlalchemy.ext.asyncio import AsyncSession
from sqlalchemy.orm import selectinload

from app.api.schemas.provider import (
    ModelCreate,
    ModelDeleteResponse,
    ModelResponse,
    ModelUpdate,
    ProviderCreate,
    ProviderDeleteResponse,
    ProviderListResponse,
    ProviderResponse,
    ProviderUpdate,
)
from app.auth.dependencies import current_user
from app.crypto.service import decrypt_api_key, encrypt_api_key
from app.db.models import Corpus, Message, Model, Provider, UsageDaily, UsageEvent, User
from app.db.session import get_session
from app.errors import BadRequest, Conflict, NotFound
from app.policy.models import WILDCARD
from app.policy.resolve import resolve_policy
from app.providers.client import normalize_base_url
from app.usage.constants import NIL_ID

router = APIRouter(
    prefix="/api/providers", tags=["providers"], dependencies=[Depends(current_user)]
)


async def _check_manage_providers(session: AsyncSession, user: User) -> bool:
    """True если admin (через *). Иначе — NotFound (не раскрываем существование)."""
    policy = await resolve_policy(session, user)
    return WILDCARD in policy.capabilities or "manage_providers" in policy.capabilities


async def _reassign_usage_daily(session: AsyncSession, model_id: str) -> None:
    """BUG-019: строки usage_daily удаляемой модели переводятся на сентинел NIL_ID.

    usage_daily хранится бессрочно (§5.3, ADR-16): агрегат обязан переживать
    удаление первичных данных, а обнулить model_id нельзя — он часть
    составного PK. Если сентинел-строка (тот же workspace/date/user) уже
    существует — счётчики мерджатся в неё. Та же семантика «записи о расходе
    сохраняются, модель отвязывается», что для message/usage_event (T-443).
    """
    rows = (
        (await session.execute(select(UsageDaily).where(UsageDaily.model_id == model_id)))
        .scalars()
        .all()
    )
    for row in rows:
        existing = await session.get(UsageDaily, (row.workspace_id, row.date, row.user_id, NIL_ID))
        if existing is not None:
            existing.requests += row.requests
            existing.tokens_in += row.tokens_in
            existing.tokens_out += row.tokens_out
            existing.cost += row.cost
            existing.errors += row.errors
            # avg_latency_ms: среднее по выборке; взвешенный пересчёт не
            # выполняется — остаётся значение существующей сентинел-строки.
            await session.delete(row)
        else:
            await session.delete(row)
            session.add(
                UsageDaily(
                    workspace_id=row.workspace_id,
                    date=row.date,
                    user_id=row.user_id,
                    model_id=NIL_ID,
                    requests=row.requests,
                    tokens_in=row.tokens_in,
                    tokens_out=row.tokens_out,
                    cost=row.cost,
                    errors=row.errors,
                    avg_latency_ms=row.avg_latency_ms,
                )
            )


def _provider_to_response(provider: Provider) -> ProviderResponse:
    return ProviderResponse(
        id=provider.id,
        kind=provider.kind,
        base_url=provider.base_url,
        enabled=provider.enabled,
        capabilities=provider.capabilities,
        models=[_model_to_response(m) for m in sorted(provider.models, key=lambda m: m.alias)],
    )


def _model_to_response(model: Model) -> ModelResponse:

    return ModelResponse(
        id=model.id,
        alias=model.alias,
        upstream_name=model.upstream_name,
        locality=model.locality,
        max_input_tokens=model.max_input_tokens,
        max_output_tokens=model.max_output_tokens,
        supports_reasoning=model.supports_reasoning,
        reasoning_toggleable=model.reasoning_toggleable,
        cost_in=model.cost_in,
        cost_out=model.cost_out,
        enabled=model.enabled,
    )


@router.post("", response_model=ProviderResponse, status_code=201)
async def create_provider(
    body: ProviderCreate,
    request: Request,
    session: AsyncSession = Depends(get_session),
    user: User = Depends(current_user),
) -> ProviderResponse:
    if not await _check_manage_providers(session, user):
        raise NotFound(
            constraint={"object": "providers", "reason": "manage_providers required"},
            hint="Нет права на управление провайдерами",
        )

    secret_key = request.app.state.secret_key

    api_key_enc = None
    if body.api_key:
        api_key_enc = encrypt_api_key(body.api_key, secret_key)

    provider = Provider(
        workspace_id=request.app.state.workspace_id,
        kind=body.kind,
        base_url=normalize_base_url(body.base_url),
        api_key_enc=api_key_enc,
        enabled=body.enabled,
        capabilities={},
    )
    session.add(provider)
    await session.commit()
    await session.refresh(provider, ["models"])

    return _provider_to_response(provider)


@router.get("", response_model=ProviderListResponse)
async def list_providers(
    request: Request,
    session: AsyncSession = Depends(get_session),
    user: User = Depends(current_user),
) -> ProviderListResponse:
    if not await _check_manage_providers(session, user):
        raise NotFound(
            constraint={"object": "providers", "reason": "manage_providers required"},
            hint="Нет права на управление провайдерами",
        )

    workspace_id = request.app.state.workspace_id
    result = await session.execute(
        select(Provider)
        .where(Provider.workspace_id == workspace_id)
        .options(selectinload(Provider.models))
    )
    providers = result.scalars().unique().all()
    return ProviderListResponse(
        providers=[_provider_to_response(p) for p in providers],
    )


@router.patch("/{provider_id}", response_model=ProviderResponse)
async def update_provider(
    provider_id: str,
    body: ProviderUpdate,
    request: Request,
    session: AsyncSession = Depends(get_session),
    user: User = Depends(current_user),
) -> ProviderResponse:
    if not await _check_manage_providers(session, user):
        raise NotFound(
            constraint={"object": "providers", "reason": "manage_providers required"},
            hint="Нет права на управление провайдерами",
        )

    workspace_id = request.app.state.workspace_id
    result = await session.execute(
        select(Provider)
        .where(Provider.id == provider_id, Provider.workspace_id == workspace_id)
        .options(selectinload(Provider.models))
    )
    provider = result.scalar_one_or_none()
    if provider is None:
        raise NotFound(
            constraint={"object": "provider", "id": provider_id},
            hint="Провайдер не найден",
        )

    if body.kind is not None:
        provider.kind = body.kind
    if body.base_url is not None:
        provider.base_url = normalize_base_url(body.base_url)
    if body.api_key is not None:
        secret_key = request.app.state.secret_key
        provider.api_key_enc = encrypt_api_key(body.api_key, secret_key)
    if body.enabled is not None:
        provider.enabled = body.enabled

    await session.commit()
    await session.refresh(provider, ["models"])

    return _provider_to_response(provider)


@router.delete("/{provider_id}", response_model=ProviderDeleteResponse)
async def delete_provider_endpoint(
    provider_id: str,
    request: Request,
    session: AsyncSession = Depends(get_session),
    user: User = Depends(current_user),
) -> ProviderDeleteResponse:
    """Удаление провайдера — только при отсутствии зарегистрированных моделей.

    Основной путь «выключить» провайдера — PATCH enabled=false. Физическое
    удаление разрешено только без моделей (заметка к T-201, T-110):
    история сообщений и расход привязаны к модели; при её удалении ссылки
    обнуляются/переносятся на сентинел (T-443, BUG-019), поэтому у
    провайдера без моделей исторических ссылок не остаётся.

    Провайдер, владеющий моделью эмбеддингов из окружения, автоматически
    блокируется этим же правилом (у него есть модель).
    """
    if not await _check_manage_providers(session, user):
        raise NotFound(
            constraint={"object": "providers", "reason": "manage_providers required"},
            hint="Нет права на управление провайдерами",
        )

    workspace_id = request.app.state.workspace_id
    result = await session.execute(
        select(Provider).where(Provider.id == provider_id, Provider.workspace_id == workspace_id)
    )
    provider = result.scalar_one_or_none()
    if provider is None:
        raise NotFound(
            constraint={"object": "provider", "id": provider_id},
            hint="Провайдер не найден",
        )

    models_count = await session.scalar(
        select(func.count()).select_from(Model).where(Model.provider_id == provider_id)
    )
    if models_count:
        raise Conflict(
            "У провайдера есть зарегистрированные модели — удаление заблокировано",
            constraint={
                "object": "provider",
                "id": provider_id,
                "reason": "has_models",
                "models_count": models_count,
            },
            hint=(
                "Удалите модели провайдера и повторите; основной путь временного "
                "отключения — кнопка «Отключить»"
            ),
        )

    await session.delete(provider)
    await session.commit()
    return ProviderDeleteResponse(deleted=True)


@router.post("/{provider_id}/probe")
async def probe_provider_endpoint(
    provider_id: str,
    request: Request,
    session: AsyncSession = Depends(get_session),
    user: User = Depends(current_user),
    deep: bool = False,
) -> dict[str, object]:
    """Запускает probe провайдера: измеряет возможности, сверяет модели.

    При deep=true — дополнительно измеряет фактический контекст (observed_context)
    для каждой available модели. Бинарный поиск, максимум 4 попытки на модель.
    Дорого: отправляет реальные промпты. Не входит в плановый ре-probe (T-112a).
    """
    if not await _check_manage_providers(session, user):
        raise NotFound(
            constraint={"object": "providers", "reason": "manage_providers required"},
            hint="Нет права на управление провайдерами",
        )

    from app.providers.probe import probe_provider

    workspace_id = request.app.state.workspace_id
    result = await session.execute(
        select(Provider)
        .where(Provider.id == provider_id, Provider.workspace_id == workspace_id)
        .options(selectinload(Provider.models))
    )
    provider = result.scalar_one_or_none()
    if provider is None:
        raise NotFound(
            constraint={"object": "provider", "id": provider_id},
            hint="Провайдер не найден",
        )

    secret_key = request.app.state.secret_key
    probe_result = await probe_provider(provider, list(provider.models), secret_key)

    provider.capabilities = {
        "available_models": probe_result.available_models,
        "supports_streaming": probe_result.supports_streaming,
        "max_parallel": probe_result.max_parallel,
        "last_probe_at": probe_result.probed_at.isoformat(),
    }
    provider.last_probe_at = probe_result.probed_at

    # T-437, часть Б: метка «уже зарегистрирована в orqion» для каждой
    # доступной модели. В capabilities (БД) остаётся чистый список имён —
    # измеренные значения; обогащение только в ответе.
    registered_names = {m.upstream_name for m in provider.models}
    response: dict[str, object] = {
        "available_models": [
            {"name": name, "registered": name in registered_names}
            for name in probe_result.available_models
        ],
        "supports_streaming": probe_result.supports_streaming,
        "max_parallel": probe_result.max_parallel,
        "model_statuses": [s.model_dump() for s in probe_result.model_statuses],
        "error": probe_result.error,
    }

    if deep and probe_result.error is None:
        from app.providers.deep_probe import measure_observed_context

        observed: dict[str, int | None] = {}
        for ms in probe_result.model_statuses:
            if ms.status == "available":
                model_obj = next(
                    (m for m in provider.models if m.id == ms.model_id),
                    None,
                )
                if model_obj is not None:
                    ctx = await measure_observed_context(provider, model_obj, secret_key)
                    observed[ms.alias] = ctx

        response["observed_context"] = observed

    await session.commit()

    return response


@router.post("/{provider_id}/models", response_model=ModelResponse, status_code=201)
async def create_model(
    provider_id: str,
    body: ModelCreate,
    request: Request,
    session: AsyncSession = Depends(get_session),
    user: User = Depends(current_user),
) -> ModelResponse:
    if not await _check_manage_providers(session, user):
        raise NotFound(
            constraint={"object": "providers", "reason": "manage_providers required"},
            hint="Нет права на управление провайдерами",
        )

    workspace_id = request.app.state.workspace_id
    result = await session.execute(
        select(Provider).where(Provider.id == provider_id, Provider.workspace_id == workspace_id)
    )
    if result.scalar_one_or_none() is None:
        raise NotFound(
            constraint={"object": "provider", "id": provider_id},
            hint="Провайдер не найден",
        )

    model = Model(
        workspace_id=workspace_id,
        provider_id=provider_id,
        alias=body.alias,
        upstream_name=body.upstream_name,
        locality=body.locality,
        max_input_tokens=body.max_input_tokens,
        max_output_tokens=body.max_output_tokens,
        supports_reasoning=body.supports_reasoning,
        reasoning_toggleable=body.reasoning_toggleable,
        cost_in=body.cost_in,
        cost_out=body.cost_out,
        enabled=body.enabled,
    )
    session.add(model)
    try:
        await session.commit()
    except IntegrityError:
        await session.rollback()
        raise BadRequest(
            "Алиас модели должен быть уникален в рамках workspace",
            hint=f"Алиас '{body.alias}' уже существует",
        )
    await session.refresh(model)

    return _model_to_response(model)


@router.patch("/models/{model_id}", response_model=ModelResponse)
async def update_model(
    model_id: str,
    body: ModelUpdate,
    request: Request,
    session: AsyncSession = Depends(get_session),
    user: User = Depends(current_user),
) -> ModelResponse:
    if not await _check_manage_providers(session, user):
        raise NotFound(
            constraint={"object": "providers", "reason": "manage_providers required"},
            hint="Нет права на управление провайдерами",
        )

    workspace_id = request.app.state.workspace_id
    result = await session.execute(
        select(Model).where(Model.id == model_id, Model.workspace_id == workspace_id)
    )
    model = result.scalar_one_or_none()
    if model is None:
        raise NotFound(
            constraint={"object": "model", "id": model_id},
            hint="Модель не найдена",
        )

    for field in (
        "alias",
        "upstream_name",
        "locality",
        "max_input_tokens",
        "max_output_tokens",
        "supports_reasoning",
        "reasoning_toggleable",
        "cost_in",
        "cost_out",
        "enabled",
    ):
        value = getattr(body, field)
        if value is not None:
            setattr(model, field, value)

    try:
        await session.commit()
    except IntegrityError:
        await session.rollback()
        raise BadRequest(
            "Алиас модели должен быть уникален в рамках workspace",
            hint=f"Алиас '{body.alias}' уже существует",
        )
    await session.refresh(model)

    return _model_to_response(model)


@router.delete("/models/{model_id}", response_model=ModelDeleteResponse)
async def delete_model(
    model_id: str,
    request: Request,
    delete_from_disk: bool = False,
    session: AsyncSession = Depends(get_session),
    user: User = Depends(current_user),
) -> ModelDeleteResponse:
    """Удаление модели провайдера — метаданные (T-443, коммит 2).

    Если модель закреплена за корпусом (corpus.pinned_model_id) — 409 без
    каскада, иначе гарантия ADR-12 терялась бы тихо. Проверка до delete:
    SELECT count(*) FROM corpus WHERE pinned_model_id = :id.

    Опциональная очистка с диска (delete_from_disk=true) — только по явному
    подтверждению из UI, не поведение по умолчанию; проводится через
    нативные эндпоинты контракта model_download.py с тем же kind-гейтом
    (DOWNLOADABLE_KINDS). Ошибка диска НЕ блокирует удаление метаданных,
    но возвращается в ответе и показывается пользователю.

    Ссылки из истории (message.model_id, usage_event.model_id) обнуляются —
    прецедент usage_event «при удалении диалога связи становятся NULL»:
    записи о сообщениях и расходе сохраняются, модель отвязывается.
    Строки бессрочного агрегата usage_daily переводятся на сентинел NIL_ID
    (BUG-019): обнуление невозможно (model_id в составе PK), удаление
    противоречило бы §5.3 — агрегат переживает первичные данные.
    """
    if not await _check_manage_providers(session, user):
        raise NotFound(
            constraint={"object": "providers", "reason": "manage_providers required"},
            hint="Нет права на управление провайдерами",
        )

    workspace_id = request.app.state.workspace_id
    result = await session.execute(
        select(Model).where(Model.id == model_id, Model.workspace_id == workspace_id)
    )
    model = result.scalar_one_or_none()
    if model is None:
        raise NotFound(
            constraint={"object": "model", "id": model_id},
            hint="Модель не найдена",
        )

    # 409: модель — чей-то пин корпуса (гарантия ADR-12)
    pinned_count = await session.scalar(
        select(func.count())
        .select_from(Corpus)
        .where(
            Corpus.pinned_model_id == model.id,
            Corpus.workspace_id == workspace_id,
        )
    )
    if pinned_count:
        raise Conflict(
            "Модель закреплена за корпусом — сначала снимите закрепление",
            constraint={"object": "model", "id": model.id, "reason": "pinned_by_corpus"},
            hint="Удалите или смените пин корпуса и повторите",
        )

    # Опциональная очистка с диска. Отложенный импорт: модельный модуль уже
    # импортирует _check_manage_providers из этого файла (цикл на уровне модуля).
    disk_deleted: bool | None = None
    disk_error: str | None = None
    if delete_from_disk:
        from app.providers.model_download import (
            DOWNLOADABLE_KINDS,
            lmstudio_delete_model,
            ollama_delete_model,
        )

        provider = await session.get(Provider, model.provider_id)
        if provider is None or provider.kind not in DOWNLOADABLE_KINDS:
            # Запрошено, но не выполнено → диск не очищен (метаданные удаляются).
            disk_deleted = False
            disk_error = (
                "Очистка с диска доступна только для локальных провайдеров (ollama, lmstudio)"
            )
        else:
            secret_key = request.app.state.secret_key
            api_key = (
                decrypt_api_key(provider.api_key_enc, secret_key) if provider.api_key_enc else None
            )
            if provider.kind == "ollama":
                disk_error = await ollama_delete_model(
                    provider.base_url, api_key, model.upstream_name
                )
            else:
                disk_error = await lmstudio_delete_model(
                    provider.base_url, api_key, model.upstream_name
                )
            disk_deleted = disk_error is None

    # Отвязываем исторические ссылки, затем удаляем саму модель.
    await session.execute(update(Message).where(Message.model_id == model.id).values(model_id=None))
    await session.execute(
        update(UsageEvent).where(UsageEvent.model_id == model.id).values(model_id=None)
    )
    await _reassign_usage_daily(session, model.id)
    await session.delete(model)
    await session.commit()

    return ModelDeleteResponse(deleted=True, disk_deleted=disk_deleted, disk_error=disk_error)