# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "orqion",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/providers": {
      "get": {
        "tags": [
          "providers"
        ],
        "summary": "List Providers",
        "operationId": "list_providers_api_providers_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProviderListResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "providers"
        ],
        "summary": "Create Provider",
        "operationId": "create_provider_api_providers_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProviderCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProviderResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/providers/{provider_id}": {
      "patch": {
        "tags": [
          "providers"
        ],
        "summary": "Update Provider",
        "operationId": "update_provider_api_providers__provider_id__patch",
        "parameters": [
          {
            "name": "provider_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Provider Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProviderUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProviderResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "providers"
        ],
        "summary": "Delete Provider Endpoint",
        "description": "Удаление провайдера — только при отсутствии зарегистрированных моделей.\n\nОсновной путь «выключить» провайдера — PATCH enabled=false. Физическое\nудаление разрешено только без моделей (заметка к T-201, T-110):\nистория сообщений и расход привязаны к модели; при её удалении ссылки\nобнуляются/переносятся на сентинел (T-443, BUG-019), поэтому у\nпровайдера без моделей исторических ссылок не остаётся.\n\nПровайдер, владеющий моделью эмбеддингов из окружения, автоматически\nблокируется этим же правилом (у него есть модель).",
        "operationId": "delete_provider_endpoint_api_providers__provider_id__delete",
        "parameters": [
          {
            "name": "provider_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Provider Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProviderDeleteResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/providers/{provider_id}/probe": {
      "post": {
        "tags": [
          "providers"
        ],
        "summary": "Probe Provider Endpoint",
        "description": "Запускает probe провайдера: измеряет возможности, сверяет модели.\n\nПри deep=true — дополнительно измеряет фактический контекст (observed_context)\nдля каждой available модели. Бинарный поиск, максимум 4 попытки на модель.\nДорого: отправляет реальные промпты. Не входит в плановый ре-probe (T-112a).",
        "operationId": "probe_provider_endpoint_api_providers__provider_id__probe_post",
        "parameters": [
          {
            "name": "provider_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Provider Id"
            }
          },
          {
            "name": "deep",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Deep"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Probe Provider Endpoint Api Providers  Provider Id  Probe Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/providers/{provider_id}/models": {
      "post": {
        "tags": [
          "providers"
        ],
        "summary": "Create Model",
        "operationId": "create_model_api_providers__provider_id__models_post",
        "parameters": [
          {
            "name": "provider_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Provider Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModelCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/providers/models/{model_id}": {
      "patch": {
        "tags": [
          "providers"
        ],
        "summary": "Update Model",
        "operationId": "update_model_api_providers_models__model_id__patch",
        "parameters": [
          {
            "name": "model_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Model Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModelUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "providers"
        ],
        "summary": "Delete Model",
        "description": "Удаление модели провайдера — метаданные (T-443, коммит 2).\n\nЕсли модель закреплена за корпусом (corpus.pinned_model_id) — 409 без\nкаскада, иначе гарантия ADR-12 терялась бы тихо. Проверка до delete:\nSELECT count(*) FROM corpus WHERE pinned_model_id = :id.\n\nОпциональная очистка с диска (delete_from_disk=true) — только по явному\nподтверждению из UI, не поведение по умолчанию; проводится через\nнативные эндпоинты контракта model_download.py с тем же kind-гейтом\n(DOWNLOADABLE_KINDS). Ошибка диска НЕ блокирует удаление метаданных,\nно возвращается в ответе и показывается пользователю.\n\nСсылки из истории (message.model_id, usage_event.model_id) обнуляются —\nпрецедент usage_event «при удалении диалога связи становятся NULL»:\nзаписи о сообщениях и расходе сохраняются, модель отвязывается.\nСтроки бессрочного агрегата usage_daily переводятся на сентинел NIL_ID\n(BUG-019): обнуление невозможно (model_id в составе PK), удаление\nпротиворечило бы §5.3 — агрегат переживает первичные данные.",
        "operationId": "delete_model_api_providers_models__model_id__delete",
        "parameters": [
          {
            "name": "model_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Model Id"
            }
          },
          {
            "name": "delete_from_disk",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Delete From Disk"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelDeleteResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ProviderCreate": {
        "properties": {
          "kind": {
            "type": "string",
            "enum": [
              "ollama",
              "lmstudio",
              "external"
            ],
            "title": "Kind"
          },
          "base_url": {
            "type": "string",
            "title": "Base Url"
          },
          "api_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Api Key"
          },
          "enabled": {
            "type": "boolean",
            "title": "Enabled",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "kind",
          "base_url"
        ],
        "title": "ProviderCreate"
      },
      "ProviderDeleteResponse": {
        "properties": {
          "deleted": {
            "type": "boolean",
            "title": "Deleted"
          }
        },
        "type": "object",
        "required": [
          "deleted"
        ],
        "title": "ProviderDeleteResponse"
      },
      "ProviderListResponse": {
        "properties": {
          "providers": {
            "items": {
              "$ref": "#/components/schemas/ProviderResponse"
            },
            "type": "array",
            "title": "Providers"
          }
        },
        "type": "object",
        "required": [
          "providers"
        ],
        "title": "ProviderListResponse"
      },
      "ProviderResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "kind": {
            "type": "string",
            "title": "Kind"
          },
          "base_url": {
            "type": "string",
            "title": "Base Url"
          },
          "enabled": {
            "type": "boolean",
            "title": "Enabled"
          },
          "capabilities": {
            "additionalProperties": true,
            "type": "object",
            "title": "Capabilities"
          },
          "models": {
            "items": {
              "$ref": "#/components/schemas/ModelResponse"
            },
            "type": "array",
            "title": "Models",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "id",
          "kind",
          "base_url",
          "enabled",
          "capabilities"
        ],
        "title": "ProviderResponse"
      },
...
    }
  }
}